Move uniparc getEntrySectionNameAndId into own util file and remove duplicate exports from types file.
